+++ components/internal/StacAsset.vue
@@ -1,18 +1,21 @@
 <template>
   <li class="vue-component asset">
     <h4>
-      <ul class="badges actions">
-        <li class="badge action download">
-          <a class="badge-fill" :href="asset.href" target="_blank" download>
-            Download '{{ asset.title || id }}'
-            <template v-if="fileFormat">as {{ fileFormat }}</template>
-          </a>
-        </li>
-      </ul>
-      <ul v-if="Array.isArray(asset.roles)" class="badges">
+      <span>{{ asset.title || id }}</span>
+      <ul v-if="Array.isArray(asset.roles)" class="badges roles">
         <li v-for="role in asset.roles" :key="role" class="badge" :class="role === 'data' ? 'green' : 'secondary'">{{ role }}</li>
       </ul>
     </h4>
+    <ul class="badges actions primary">
+      <li class="badge action download">
+        <a class="badge-fill" :href="asset.href" target="_blank" download>
+          <span class="icon">💾</span> Download <template v-if="fileFormat">{{ fileFormat }}</template>
+        </a>
+      </li>
+      <li class="badge action copy" @click.prevent.stop="copyURL($event, asset.href)">
+        <span class="icon">📋</span> Copy URL
+      </li>
+    </ul>
     <Description v-if="asset.description" :description="asset.description" :compact="true" />
     <StacFields type="Asset" :metadata="asset" :ignore="ignore" title="" :context="context" headingTag="h5" />
   </li>
@@ -22,6 +25,7 @@
 import { Formatters } from '@radiantearth/stac-fields';
 import Description from '../Description.vue';
 import StacFields from './StacFields.vue';
+import CopyMixin from './CopyMixin';
 
 export default {
   name: 'Asset',
@@ -29,6 +33,7 @@ export default {
     Description,
     StacFields
   },
+  mixins: [CopyMixin],
   props: {
     asset: {
       type: Object,
@@ -55,25 +60,42 @@ export default {
       }
       return null;
     }
+  },
+  methods: {
+		copyURL(event, url) {
+      const elem = event.composedPath()[0].querySelector('.icon'); // event.target doesn't work in web components
+      this.copyText(url, () => this.toggleIcon(elem, '✅'), () => this.toggleIcon(elem, '❌'));
+		}
   }
 }
 </script>
 
 <style lang="scss">
 .vue-component.asset {
+  margin-bottom: 1.5em;
+
+  > * {
+    margin: 0.25em 0;
+  }
+
+  .actions > .action {
+    padding: 0.5em;
+  }
+
   h4  {
     font-size: 1.1em;
-    margin: 0;
     font-weight: normal;
 
-    .download {
-      margin-right: 1em;
+    > .roles {
+      font-size: 0.95em;
+      margin-left: 0.3em;
     }
   }
+
   h5  {
     font-size: 1.1em;
     font-weight: bold;
-    margin: 0.75em 0 0.5em;
+    margin: 0.5em 0;
   }
 
   .metadata {